Forbid having both read and write locks in a Guard
The previous formulation of Guard allowed (in the types, at least)
having Some(_) for both r_guard and w_guard. But this should never be
allowed.

It in fact was not ever constructed in this manner, and the fields are
private so we know it never will be. But let's forbid it with types.

Not totally sure if the drop impl is right, I'm not sure what reading
the r_guard did in the previous impl.