v0.2
This is a significant rewrite of xshell. The main change is the switch
from global-state API to object based API -- `Shell` is now a thing and
is the gateway to all API.

Motivation is manyfold:

* Curren `fs` API is unergonomic, as it requires importing many
  individual functions. With the new style API, `use xshell::{Shell,
  cmd};` is all you need, and API is discoverable via `sh.`
  auto-completion.
* Soundness. The recent discussion on IRLO
  <https://internals.rust-lang.org/t/synchronized-ffi-access-to-posix-environment-variable-functions/15475>
  convinced me that process' environment shold be read-only
* Implementation complexity: we spend a lot of effort debugging various
  issues related to `gsl`.

Specific changes:

* There's now a `Shell` which holds the environment. For ergonomics
  reasons, `Shell` uses interior mutability.
* The `cmd!` macro now requires a `Shell` as a first argument.
* The `Cmd` builder methods are somewhat more orthogonal.
* The suggested naming no longer follows UNIXism which makes the naming
  more incluiseve and natural (assuming you have code completion).

With the present API, I also feel we should shift from "convenient
hacks" to "robust tool". As a first step to that end, stdin is now fed
via a dedicated thread, to aviod deadlock.